Add Duration and Effect Stacking Flags
Note: This change introduces `void ApplyEffect` to TimedBuff. This
method should be overridden and buff effect logic should be implemented
here. `Apply()` is not longer expected to be overridden and is used for
timedBuff logic. This is a breaking change but I feel that this is the
best approach in regards to the tutorial.

* Duration flag determines if buff time should be extended on each
application of the buff.
* Effect stacking flag determines if the buff value should be
incremented on each application of the buff.

* Added MovementComponent.cs for clarity
* Refactored various variables and naming